requirements.txt
discord.py
fandom-py
beautifulsoup4
python-dotenv
requests
